MGE XE 0.9.3
Fixed crash when turning off distant land, due to trying to access
parameters for an interior when in an exterior.

Altered distant land code to reset fog and the clipping plane to
Morrowind defaults when it is toggled.

Made shadows render properly on CCW culled objects. They are uncommon but
can be generated by the engine when mirroring body parts.

HDR shader exposure made brighter; it was lowering overall brightness in
too many scenes.