0.9.8 bugfix
Serious failure to assign MGEXEgui localization properly at initialization
time fixed.

Problems with incorrect bumpmap behaviour of FFE fixed. It was a state
tracking issue. It was a bad idea to assume that there would be no spurious
state updates for non-bumpmapped objects, and the bump data comes from the
first stage of two, not the texgen stage. Texgen is also used by itself
for reflections.