Shadow atlas update
Restricted shadow map rendering to the area inside the view frustum, by rendering the frustum into the stencil then using the stencil test to reject pixels outside it.

While it does not improve the worst case when the view is parallel to the sun, it can improve performance in the average case.