Distant land UI update
Created a better checked list control to improve the poor UX of the Windows Forms version. The old version was click to select, then click again to toggle the check. Now, clicking within the check box toggles on the first click, and double clicking anywhere with the item also toggles. Based on a post from the Windows dev forums.