Eliminate shadow artifacts in per-pixel lighting mode, and reduce artifacts in standard lighting mode. Turns off shadow biasing when per-pixel lighting is active, and adjusts the bias in standard lighting.
Extra dark shadows can occur where there are overlapping meshes with surfaces that are very co-planar, which are most visible in overlapping elbows and wrists of body models. This ultimately is caused by a mismatch between fixed function and shader rendering. MGE XE has to apply a depth bias to avoid Z-fighting, but this also causes overlapping multiple shadow rendering when Z is nearly equal. This can be fixed by adjusting the models correctly, but they are designed foremost for Morrowind's body part system that turns off parts individually.

This commit addresses issue #7.